requirements.txt
dlib~=19.24.0
pillow~=9.4.0
pyside6~=6.4.2
numpy~=1.24.2
filetype~=1.2.0
qtawesome~=1.3.0.dev0
